// src/app/shared/models/cotizacion.model.ts
// Generated by https://quicktype.io

export interface Cotizacion {
    id:                 number;
    correlativo?:      string;
    descuentoAdicional?: number;
    diasCotizacion:        number;
    fecha?:              string;
    descripcion?:              string;

    idCliente?:          number;
    nombreCliente?:      string;
    codigoCliente?:      string;
    emailCliente?:      string;

    idSucursal?:         number;
    numeroSucursal?:         number;

    fechaLimite?:              string;
    total?:              number;
    totalSujetoIva?:     number;
    detalle?:            CotizacionDetalle[];
    usuario?:              string;

    idEmpresa?:          number;
    nit?:          number;
}

export interface CotizacionDetalle {
    cantidad:        number;
    codigoProducto?:  string;
    descuento:       number;
    id?:              number;
    idProducto?:      number;
    idTipoDescuento?: number;
    idTipoProducto?:  number;
    idCotizacion?:         number;
    precio:          number;
    producto?:        string;
    subtotal:        number;
    total:           number;
    valorDescuento?:  number;
    codigo?:     string;
    codigoStock?:     string;
    descripcionAdicional?:     string;
}