// src/app/shared/models/venta.model.ts
// Generated by https://quicktype.io

import { Cliente } from "./cliente.model";
import { Pago } from "./pago.model";

export interface Venta {
    correlativo?:      string;
    descuentoAdicional: number;
    diasCredito:        number;
    estadoVenta?:        string;
    fecha?:              string;
    id:                 number;

    //cliente?:               any;
    idCliente?:          number;
    nombreCliente?:      string;
    codigoCliente?:      string;
    emailCliente?:      string;

    idEstadoVenta:      number;
    idSucursal:         number;
    numeroSucursal?:         number;
    idTipoVenta:        number;

    tipoVenta?:          string;
    total:              number;
    totalSujetoIva:     number;
    saldo?:              number;
    fechaLimite?:              string;
    itemsEliminados?:    number[]|null;
    detalle?:            VentaDetalle[];
    factura?:           VentaFacturaResumen;
    usuario?:              string;
    idMesa?:      number|null;
    mesa?:              string|null;
}

export interface VentaDetalle {
    cantidad:        number;
    codigoProducto?:  string;
    descuento:       number;
    id?:              number;
    idProducto?:      number;
    idTipoDescuento?: number;
    idTipoProducto?:  number;
    idVenta?:         number;
    numeroImei?:      string;
    numeroSerie?:     string;
    observacion?:     string;
    precio:          number;
    producto?:        string;
    subtotal:        number;
    total:           number;
    valorDescuento?:  number;
    codigo?:     string;
    codigoStock?:     string;
    descripcionAdicional?:     string;
}

export interface VentaFacturaResumen {
    id:                number;
    codigoRecepcion:  string;
    numeroFactura   :  number;
    cuf:              string;
    codigoEstado:     number;
    estado:           string;
    codigoAsociacion:          string;
    idVenta:          string;
    nitEmisor   :  number;
}

export interface SolicitudGeneracionQr {
    idVenta:      number;
    generarNuevo?:  boolean;

}


export interface SolicitudVentaOnline {
    cliente:      ClienteOnline;
    venta: VentaOnline;
    remision:        boolean;
    idEmpresa:     number;
}

export interface VentaOnline {
    total:              number;
    descuentoAdicional: number;
    totalSujetoIva:     number;
    idSucursal:         number;
    detalle:            VentaDetalle[];
    usuario:              string;
    codigoTipoPago:     number
}

export interface ClienteOnline {
    codigoCliente: string;
    codigoTipoDocumentoIdentidad: number;
    numeroDocumento: string;
    complemento: string;
    nombre: string;
    direccion: string;
  }